feat: renombrados packages, excepciones e interfaces

+++ src/es/cursojava/oo/excepciones/ejercicios/ejercicio3/pojos/TazaCafe.java
@@ -2,9 +2,9 @@ package es.cursojava.oo.excepciones.ejercicios.ejercicio3.pojos;
 
 public class TazaCafe {
 	private String tipo;
-	private float temperatura;
+	private double temperatura;
 	
-	public TazaCafe(String tipo, float temperatura) {
+	public TazaCafe(String tipo, double temperatura) {
 		super();
 		this.tipo = tipo;
 		this.temperatura = temperatura;
@@ -18,11 +18,11 @@ public class TazaCafe {
 		this.tipo = tipo;
 	}
 	
-	public float getTemperatura() {
+	public double getTemperatura() {
 		return temperatura;
 	}
 	
-	public void setTemperatura(float temperatura) {
+	public void setTemperatura(double temperatura) {
 		this.temperatura = temperatura;
 	}
 }
